+++ scripts/dev.sh
@@ -1,6 +1,11 @@
 tmux \
   new$([ -n "$TMUX" ] && echo -n w) \; \
   set mouse on \; \
+  renamew '📝 docs:serve' \; \
+  send 'npx serve _docs' C-m \; \
+  splitw -h \; \
+  send 'npm run docs:watch' C-m \; \
+  neww \; \
   renamew '🔨 dev' \; \
   send 'npm run watch' C-m \; \
   splitw \; \