+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add mouse support for scripts that use tmux
 - Set window name for scripts that use tmux
 - Support scripts using tmux in tmux
+- Make `./scripts/dev.sh` include content of `./scripts/docs-serve.sh`
 
 ## [0.3.0]
 ## Added